Safari-609.1.11.1/LayoutTests/accessibility/media-with-aria-label.html

<!DOCTYPE HTML><!-- webkit-test-runner [ enableModernMediaControls=false ] -->
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>

<video id="video1" width=100 height=100 controls aria-label="Video One"></video>

<br>

<audio id="audio1" width=100 height=100 controls aria-label="Audio One"></audio>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests ensures that aria-label is conveyed for video and audio elements.")

    if (window.accessibilityController) {
        var video1 = accessibilityController.accessibleElementById("video1");
        shouldBe("video1.description", "'AXDescription: Video One'");
        var audio1 = accessibilityController.accessibleElementById("audio1");
        shouldBe("audio1.description", "'AXDescription: Audio One'");
    }
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
